Reorganized geometry.warper package
