+++ kornia/geometry/__init__.py
@@ -2,7 +2,6 @@ from kornia.geometry.camera import *
 from kornia.geometry.conversions import *
 from kornia.geometry.linalg import *
 from kornia.geometry.transform import *
-from kornia.geometry.warp import *
 from kornia.geometry.depth import *
 from kornia.geometry.subpix import *
 from kornia.geometry.epipolar import *
