+++ kornia/geometry/transform/__init__.py
@@ -7,3 +7,4 @@ from kornia.geometry.transform.affwarp import *
 from kornia.geometry.transform.flips import *
 from kornia.geometry.transform.projwarp import *
 from kornia.geometry.transform.elastic_transform import *
+from kornia.geometry.transform.homography_warper import *
